Safari-611.1.10/LayoutTests/svg/custom/use-display-none.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">
    <symbol id="a"/>
    <use id="b" xlink:href="#a"/>
    <script>
        if (window.testRunner)
            testRunner.dumpAsText();
        b=document.getElementById('b');
        b.href.baseVal="#a";
        b.setAttribute("display","none");
    </script>
    <text y="20">PASS: The text is displayed and the file did not trigger a crash.</text>
</svg>
