third_party/blink/web_tests/platform/generic/svg/custom/use-display-none-expected.txt

PASS: The text is displayed and the file did not trigger a crash.
